Moving reflection to/from into appropriate read/write headers
