+++ include/glaze/glaze.hpp
@@ -37,7 +37,6 @@
 #include "glaze/file/file_ops.hpp"
 #include "glaze/json.hpp"
 #include "glaze/record/recorder.hpp"
-#include "glaze/reflection/reflect.hpp"
 
 namespace glz
 {
