+++ include/glaze/json/read.hpp
@@ -23,6 +23,7 @@
 #include "glaze/util/strod.hpp"
 #include "glaze/util/type_traits.hpp"
 #include "glaze/util/variant.hpp"
+#include "glaze/reflection/reflect.hpp"
 
 namespace glz
 {
@@ -2430,6 +2431,161 @@ namespace glz
                   }
                }
                read<json>::op<Opts>(*value, ctx, it, end);
+            }
+         }
+      };
+      
+      template <reflectable T>
+      struct from_json<T>
+      {
+         template <auto Options, string_literal tag = "">
+         GLZ_FLATTEN static void op(auto&& value, is_context auto&& ctx, auto&& it, auto&& end)
+         {
+            parse_object_opening<Options>(ctx, it, end);
+            if (bool(ctx.error)) [[unlikely]]
+               return;
+
+            static constexpr auto Opts = opening_handled_off<ws_handled_off<Options>()>();
+
+            static constexpr auto num_members = std::tuple_size_v<decltype(to_tuple(std::declval<T>()))>;
+
+            if constexpr (num_members == 0 && Options.error_on_unknown_keys) {
+               if (*it == '}') [[likely]] {
+                  ++it;
+                  return;
+               }
+               ctx.error = error_code::unknown_key;
+               return;
+            }
+            else {
+               // Only used if error_on_missing_keys = true
+               [[maybe_unused]] bit_array<num_members> fields{};
+
+               static constinit auto frozen_map = make_reflection_map<T, Opts.use_hash_comparison>();
+               // we have to populate the pointers in the reflection map from the structured binding
+               auto t = to_tuple(value);
+               for_each<num_members>([&](auto I) {
+                  std::get<std::add_pointer_t<std::decay_t<decltype(std::get<I>(t))>>>(
+                     std::get<I>(frozen_map.items).second) = &std::get<I>(t);
+               });
+
+               bool first = true;
+               while (true) {
+                  if (*it == '}') [[unlikely]] {
+                     ++it;
+                     if constexpr (Opts.error_on_missing_keys) {
+                        constexpr auto req_fields = required_fields<T, Opts>();
+                        if ((req_fields & fields) != req_fields) {
+                           ctx.error = error_code::missing_key;
+                        }
+                     }
+                     return;
+                  }
+                  else if (first) [[unlikely]]
+                     first = false;
+                  else [[likely]] {
+                     match<','>(ctx, it, end);
+                     if (bool(ctx.error)) [[unlikely]]
+                        return;
+                     skip_ws_no_pre_check<Opts>(ctx, it, end);
+                     if (bool(ctx.error)) [[unlikely]]
+                        return;
+                  }
+
+                  if constexpr (num_members == 0) {
+                     // parsing to an empty object, but at this point the JSON presents keys
+                     const sv key = parse_object_key<T, ws_handled<Opts>(), tag>(ctx, it, end);
+                     if (bool(ctx.error)) [[unlikely]]
+                        return;
+
+                     if constexpr (Opts.error_on_unknown_keys) {
+                        if constexpr (tag.sv().empty()) {
+                           std::advance(it, -int64_t(key.size()));
+                           ctx.error = error_code::unknown_key;
+                           return;
+                        }
+                        else if (key != tag.sv()) {
+                           std::advance(it, -int64_t(key.size()));
+                           ctx.error = error_code::unknown_key;
+                           return;
+                        }
+                     }
+                     else {
+                        parse_object_entry_sep<Opts>(ctx, it, end);
+                        if (bool(ctx.error)) [[unlikely]]
+                           return;
+
+                        skip_value<Opts>(ctx, it, end);
+                        if (bool(ctx.error)) [[unlikely]]
+                           return;
+                     }
+                  }
+                  else {
+                     const sv key = parse_object_key<T, ws_handled<Opts>(), tag>(ctx, it, end);
+                     if (bool(ctx.error)) [[unlikely]]
+                        return;
+
+                     // Because parse_object_key does not necessarily return a valid JSON key, the logic for handling
+                     // whitespace and the colon must run after checking if the key exists
+                     if (const auto& member_it = frozen_map.find(key); member_it != frozen_map.end()) [[likely]] {
+                        parse_object_entry_sep<Opts>(ctx, it, end);
+                        if (bool(ctx.error)) [[unlikely]]
+                           return;
+
+                        if constexpr (Opts.error_on_missing_keys) {
+                           // TODO: Kludge/hack. Should work but could easily cause memory issues with small changes.
+                           // At the very least if we are going to do this add a get_index method to the maps and call
+                           // that
+                           auto index = member_it - frozen_map.begin();
+                           fields[index] = true;
+                        }
+                        std::visit(
+                           [&](auto&& member_ptr) {
+                              read<json>::op<ws_handled<Opts>()>(get_member(value, member_ptr), ctx, it, end);
+                           },
+                           member_it->second);
+                        if (bool(ctx.error)) [[unlikely]]
+                           return;
+                     }
+                     else [[unlikely]] {
+                        if constexpr (Opts.error_on_unknown_keys) {
+                           if constexpr (tag.sv().empty()) {
+                              std::advance(it, -int64_t(key.size()));
+                              ctx.error = error_code::unknown_key;
+                              return;
+                           }
+                           else if (key != tag.sv()) {
+                              std::advance(it, -int64_t(key.size()));
+                              ctx.error = error_code::unknown_key;
+                              return;
+                           }
+                           else {
+                              // We duplicate this code to avoid generating unreachable code
+                              parse_object_entry_sep<Opts>(ctx, it, end);
+                              if (bool(ctx.error)) [[unlikely]]
+                                 return;
+
+                              skip_value<Opts>(ctx, it, end);
+                              if (bool(ctx.error)) [[unlikely]]
+                                 return;
+                           }
+                        }
+                        else {
+                           // We duplicate this code to avoid generating unreachable code
+                           parse_object_entry_sep<Opts>(ctx, it, end);
+                           if (bool(ctx.error)) [[unlikely]]
+                              return;
+
+                           skip_value<Opts>(ctx, it, end);
+                           if (bool(ctx.error)) [[unlikely]]
+                              return;
+                        }
+                     }
+                  }
+                  skip_ws<Opts>(ctx, it, end);
+                  if (bool(ctx.error)) [[unlikely]]
+                     return;
+               }
             }
          }
       };
